+++ db/hooks.php
@@ -26,6 +26,10 @@ defined('MOODLE_INTERNAL') || die();
 
 $callbacks = [
     [
+        'hook' => \core\hook\di_configuration::class,
+        'callback' => \aiprovider_router\hook_listener::class . '::configure_di',
+    ],
+    [
         'hook' => \core_ai\hook\after_ai_provider_form_hook::class,
         'callback' => \aiprovider_router\hook_listener::class . '::set_form_definition_for_aiprovider_router',
     ],